Moved circular deque into its own sub module
